Extract app descriptor value types
They will be used in the link descriptor.
